+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- Support Amplify configurations with phone number as username
+- Added support for several required attributes, including address, name, website and many more
+
 ## 0.1.0
 
 - Initial version